+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.2.0 (Next)
 
 * Replaces non-breaking spaces (`&nbsp;`) with regular spaces in converted Markdown - [@dblock](https://github.com/dblock).
+* Your contribution here.
 
 ### 0.1.0 (2026/09/19)